Converted banning modal into a reuseable directive
* Created ban-modal directive to remove a lot of duplicate code

* Management->Users, Moderation->Users,Posts, and Messages, and the
  Profile page now all use the same banning directive
