+++ client/admin/moderation/messages.html
@@ -196,86 +196,7 @@
 
 
 <!-- Manage Bans Modal -->
-<modal show="ModerationCtrl.showManageBansModal" on-close="ModerationCtrl.closeManageBans()">
-  <form name="$parent.form" class="css-form">
-    <h3 class="thin-underline">Manage Bans</h3>
-    <label class="bold">User Information</label>
-    <table class="striped" width="100%">
-      <tr>
-        <td>Username</td>
-        <td ng-bind="ModerationCtrl.selectedUser.username"></td>
-      </tr>
-      <tr>
-        <td>Email</td>
-        <td ng-bind="ModerationCtrl.selectedUser.email"></td>
-      </tr>
-      <tr>
-        <td>Register Date</td>
-        <td ng-bind="ModerationCtrl.selectedUser.created_at | humanDate"></td>
-      </tr>
-      <tr ng-if="ModerationCtrl.selectedUser.ban_expiration">
-        <td>Global Ban Expiration</td>
-        <td ng-bind="ModerationCtrl.selectedUser.ban_expiration | humanDate:true"></td>
-      </tr>
-      <tr ng-if="ModerationCtrl.selectedUser.banned_board_names.length">
-        <td>Banned From Boards</td>
-        <td>
-          <span ng-repeat="boardName in ModerationCtrl.selectedUser.banned_board_names track by $index" ng-bind-html="$last ? boardName : boardName + ', '"></span>
-        </td>
-      </tr>
-    </table>
-
-    <div ng-if="ModerationCtrl.canGlobalBanUser()">
-      <label for="banType">
-        <strong>Ban Globally</strong>
-        <a class="right" ng-click="ModerationCtrl.permanentBan = undefined; ModerationCtrl.banUntil = undefined;" ng-show="ModerationCtrl.permanentBan !== undefined">
-          <i class="fa fa-times"></i> Remove Global Ban
-        </a>
-      </label>
-      <input type="radio" name="banType" ng-model="ModerationCtrl.permanentBan" ng-value="true" id="permanent" ng-disabled="ModerationCtrl.banSubmitted"><label for="permanent">Permanent</label>
-      <input type="radio" name="banType" ng-model="ModerationCtrl.permanentBan" ng-value="false" id="temporary" ng-disabled="ModerationCtrl.banSubmitted"><label for="temporary">Temporary</label>
-
-      <div ng-show="ModerationCtrl.permanentBan === false">
-        <label for="banUntil">Enter Ban Expiration Date:</label>
-        <input type="date" ng-model="ModerationCtrl.banUntil" min="{{ModerationCtrl.minDate()}}" name="banUntil" ng-required="ModerationCtrl.permanentBan === false" />
-      </div>
-
-      <div ng-show="ModerationCtrl.showIpBan && ModerationCtrl.permanentBan === true">
-        <input type="checkbox" ng-model="ModerationCtrl.banUserIp" ng-disabled="ModerationCtrl.banSubmitted" name="ipBan" id="ipBan" /><label for="ipBan">Ban all of {{ModerationCtrl.selectedUser.username}}'s known IP addresses</label>
-      </div>
-    </div>
-
-    <label>
-      <strong>Ban from my Boards</strong>
-      <div class="right">
-        <a ng-href="#" ng-click="ModerationCtrl.checkModBoards();"><i class="fa fa-check-square-o"></i> Check All</a>
-        &nbsp;&nbsp;
-        <a ng-href="#" ng-click="ModerationCtrl.uncheckModBoards();"><i class="fa fa-square-o"></i> Uncheck All</a>
-      </div>
-    </label>
-    <div class="board-bans-list">
-      <script type="text/ng-template" id="board_renderer.html">
-        <input ng-init="ModerationCtrl.loadBoardBans(board.id)" id="{{board.id}}" ng-checked="ModerationCtrl.boardBanList.indexOf(board.id) > -1" value="{{board.id}}" ng-click="ModerationCtrl.toggleBoardBan(board.id)" ng-disabled="!ModerationCtrl.canBoardBanUser(board.id)" type="checkbox" />
-        <label for="{{board.id}}" ng-bind-html="board.name"></label>
-        <ul>
-          <li ng-repeat="board in board.children track by board.id" ng-include="'board_renderer.html'"></li>
-        </ul>
-      </script>
-      <div ng-repeat="cat in ModerationCtrl.boards track by cat.id">
-        <label class="bold" ng-bind-html="cat.name"></label>
-        <ul>
-          <li ng-repeat="board in cat.boards track by board.id" ng-include="'board_renderer.html'">
-          </li>
-        </ul>
-      </div>
-    </div>
-
-    <div class="clear">
-      <button class="fill-row" ng-click="ModerationCtrl.updateBans()"
-        ng-disabled="$parent.form.$invalid || $parent.form.$errors || ModerationCtrl.banSubmitted" ng-bind="ModerationCtrl.confirmBanBtnLabel"></button>
-    </div>
-  </form>
-</modal>
+<ban-modal show-modal="ModerationCtrl.showManageBansModal" selected-user="ModerationCtrl.selectedUser" callback="ModerationCtrl.updateReportBans"></ban-modal>
 
 <!-- Set Status Modal -->
 <modal show="ModerationCtrl.showSetStatusModal" on-close="ModerationCtrl.closeSetStatus()">
