+++ client/components/profile/profile.html
@@ -405,86 +405,7 @@
 </modal>
 
 <!-- Manage Bans Modal -->
-<modal show="vm.showManageBansModal" on-close="vm.closeManageBans()">
-  <form name="$parent.form" class="css-form">
-    <h3 class="thin-underline">Manage Bans</h3>
-    <label class="bold">User Information</label>
-    <table class="striped" width="100%">
-      <tr>
-        <td>Username</td>
-        <td ng-bind="vm.selectedUser.username"></td>
-      </tr>
-      <tr>
-        <td>Email</td>
-        <td ng-bind="vm.selectedUser.email"></td>
-      </tr>
-      <tr>
-        <td>Register Date</td>
-        <td ng-bind="vm.selectedUser.created_at | humanDate"></td>
-      </tr>
-      <tr ng-if="vm.selectedUser.ban_expiration">
-        <td>Global Ban Expiration</td>
-        <td ng-bind="vm.selectedUser.ban_expiration | humanDate:true"></td>
-      </tr>
-      <tr ng-if="vm.selectedUser.banned_board_names.length">
-        <td>Banned From Boards</td>
-        <td>
-          <span ng-repeat="boardName in vm.selectedUser.banned_board_names track by $index" ng-bind-html="$last ? boardName : boardName + ', '"></span>
-        </td>
-      </tr>
-    </table>
-
-    <div ng-if="vm.canGlobalBanUser()">
-      <label for="banType">
-        <strong>Ban Globally</strong>
-        <a class="right" ng-click="vm.permanentBan = undefined; vm.banUntil = undefined;" ng-show="vm.permanentBan !== undefined">
-          <i class="fa fa-times"></i> Remove Global Ban
-        </a>
-      </label>
-      <input type="radio" name="banType" ng-model="vm.permanentBan" ng-value="true" id="permanent" ng-disabled="vm.banSubmitted"><label for="permanent">Permanent</label>
-      <input type="radio" name="banType" ng-model="vm.permanentBan" ng-value="false" id="temporary" ng-disabled="vm.banSubmitted"><label for="temporary">Temporary</label>
-
-      <div ng-show="vm.permanentBan === false">
-        <label for="banUntil">Enter Ban Expiration Date:</label>
-        <input type="date" ng-model="vm.banUntil" min="{{vm.minDate()}}" name="banUntil" ng-required="vm.permanentBan === false" />
-      </div>
-
-      <div ng-show="vm.showIpBan && vm.permanentBan === true">
-        <input type="checkbox" ng-model="vm.banUserIp" ng-disabled="vm.banSubmitted" name="ipBan" id="ipBan" /><label for="ipBan">Ban all of {{vm.selectedUser.username}}'s known IP addresses</label>
-      </div>
-    </div>
-
-    <label>
-      <strong>Ban from my Boards</strong>
-      <div class="right">
-        <a ng-href="#" ng-click="vm.checkModBoards();"><i class="fa fa-check-square-o"></i> Check All</a>
-        &nbsp;&nbsp;
-        <a ng-href="#" ng-click="vm.uncheckModBoards();"><i class="fa fa-square-o"></i> Uncheck All</a>
-      </div>
-    </label>
-    <div class="board-bans-list">
-      <script type="text/ng-template" id="board_renderer.html">
-        <input ng-init="vm.loadBoardBans(board.id)" id="{{board.id}}" ng-checked="vm.boardBanList.indexOf(board.id) > -1" value="{{board.id}}" ng-click="vm.toggleBoardBan(board.id)" ng-disabled="!vm.canBoardBanUser(board.id)" type="checkbox" />
-        <label for="{{board.id}}" ng-bind-html="board.name"></label>
-        <ul>
-          <li ng-repeat="board in board.children track by board.id" ng-include="'board_renderer.html'"></li>
-        </ul>
-      </script>
-      <div ng-repeat="cat in vm.boards track by cat.id">
-        <label class="bold" ng-bind-html="cat.name"></label>
-        <ul>
-          <li ng-repeat="board in cat.boards track by board.id" ng-include="'board_renderer.html'">
-          </li>
-        </ul>
-      </div>
-    </div>
-
-    <div class="clear">
-      <button class="fill-row" ng-click="vm.updateBans()"
-        ng-disabled="$parent.form.$invalid || $parent.form.$errors || vm.banSubmitted" ng-bind="vm.confirmBanBtnLabel"></button>
-    </div>
-  </form>
-</modal>
+<ban-modal show-modal="vm.showBanModal" selected-user="vm.selectedUser" callback="vm.updateUser"></ban-modal>
 
 <!-- Edit Preferences Modal -->
 <modal class="medium" show="vm.editPreferences">
